+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+## Added
+ - Endpoints to get RabbitMQ queue details about messages and consumers for extractions and conversions. 
+ [BD-2184](https://opensource.ncsa.illinois.edu/jira/browse/BD-2184)
+
 ## 0.4.0 - 2018-05-15
 
 ### Changed